API redesign: enable custom serializers & deserializers
