+++ pom.xml
@@ -11,7 +11,7 @@
     <inceptionYear>2014</inceptionYear>
 
     <packaging>bundle</packaging>
-    <version>2.2.1-SNAPSHOT</version>
+    <version>2.3.0-SNAPSHOT</version>
 
     <organization>
         <name>exxeleron</name>
